Add files via upload
UnsupervisedLearningPCA_VehicleSilhoutte_Classification.ipynb

DESCRIPTION:
The data contains features extracted from the silhouette of vehicles in different angles. Four "Corgie" model vehicles were used for the experiment: a double decker bus, Cheverolet van, Saab 9000 and an Opel Manta 400 cars. This particular combination of vehicles was chosen with the expectation that the bus, van and either one of the cars would be readily distinguishable, but it would be more difficult to distinguish between the cars.

DOMAIN: Object recognition

CONTEXT:
The purpose is to classify a given silhouette as one of three types of vehicle, using a set of features extracted from the silhouette. The vehicle may be
viewed from one of many different angles.
Learning Outcomes:
● Exploratory Data Analysis
● Reduce number dimensions in the dataset with minimal information loss
● Train a model using Principal Components

OBJECTIVE:
Apply dimensionality reduction technique – PCA and train a model using
principal components instead of training the model using just the raw data.